+++ src/common/document.js
@@ -1,5 +1,3 @@
-import { addConfigurationRow, saveConfiguration } from "../options/helper";
-
 let doc; 
 
 export function initializeDocument(injectedDocument) {
@@ -14,14 +12,6 @@ export function updateJiraStatusStyle(element, configuration) {
     element.style.backgroundColor = configuration.backgroundColour;
     element.style.color = configuration.textColour;
     return element;
-}
-
-export function initializeEventListenersForOptions() {
-  doc.getElementById("SaveConfiguration")
-    .addEventListener("click", saveConfiguration);
-  
-  doc.getElementById("AddConfigurationRow")
-    .addEventListener("click", addConfigurationRow);
 }
 
 export function getElementById(id) {
