+++ src/options/componentHelper.js
