+++ src/options/index.js
@@ -1,7 +1,8 @@
-import { initializeDocument, getElementById, initializeEventListenersForOptions } from "../common/document";
+import { initializeDocument, getElementById } from "../common/document";
 import { presetConfiguration, rowCount, chromeSyncStorageKey } from "../common/settings";
 import { getStorage } from "../common/storage";
-import { createRow } from "./documentHelper";
+import { createRow } from "./componentHelper";
+import { addConfigurationRow, saveConfiguration } from "./helper";
 
 export function initialLoad(result) {
     const savedConfiguration = result || presetConfiguration;        
@@ -11,6 +12,11 @@ export function initialLoad(result) {
         rowCount++;
         createRow(rowCount, parentElement, savedConfiguration[key], key);
     }
+}
+
+export function initializeEventListenersForOptions() {
+    getElementById("SaveConfiguration").addEventListener("click", saveConfiguration);
+    getElementById("AddConfigurationRow").addEventListener("click", addConfigurationRow);
 }
 
 window.onload = function() {
