+++ src/tests/mocks/documentMock.js
@@ -1,15 +1,13 @@
 export const getElementById = jest.fn();
-export const getElements = jest.fn();
-export const querySelectorAll = jest.fn();
+export const updateJiraStatusStyle = jest.fn();
 export const createElement = jest.fn();
 export const appendChild = jest.fn();
-export const prepend = jest.fn();
+export const getJiraStatusLabels = jest.fn();
 
 export const document = {
     getElementById,
-    getElements,
-    querySelectorAll,
+    updateJiraStatusStyle,
     createElement,
     appendChild,
-    prepend
+    getJiraStatusLabels
 };
